organise functions in deriv-data-processor
