+++ services/deriv-data-processor/utils.js
@@ -1,0 +1,12 @@
+// change tick format
+const changeTickFormat = (tick) => {
+  return {
+    symbol: tick.symbol,
+    date: new Date(tick.epoch * 1000),
+    price: tick.quote,
+  };
+};
+
+module.exports = {
+  changeTickFormat,
+};
